+++ tableauserverclient/server/request_options.py
@@ -11,6 +11,7 @@ class RequestOptions(RequestOptionsBase):
         LessThan = 'lt'
         LessThanOrEqual = 'lte'
         In = 'in'
+        Has = 'has'
 
     class Field:
         Args = 'args'
@@ -82,6 +83,7 @@ class RequestOptions(RequestOptionsBase):
 
 class _FilterOptionsBase(RequestOptionsBase):
     """ Provide a basic implementation of adding view filters to the url """
+
     def __init__(self):
         self.view_filters = []