+++ tableauserverclient/server/server.py
@@ -3,8 +3,8 @@ import xml.etree.ElementTree as ET
 from .exceptions import NotSignedInError
 from ..namespace import Namespace
 from .endpoint import Sites, Views, Users, Groups, Workbooks, Datasources, Projects, Auth, \
-    Schedules, ServerInfo, Tasks, ServerInfoEndpointNotFoundError, Subscriptions, Jobs, Metadata,\
-    Databases, Tables, Flows, Webhooks
+    Schedules, ServerInfo, Tasks, Subscriptions, Jobs, Metadata,\
+    Databases, Tables, Flows, Webhooks, DataAccelerationReport
 from .endpoint.exceptions import EndpointUnavailableError, ServerInfoEndpointNotFoundError
 
 import requests
@@ -56,6 +56,7 @@ class Server(object):
         self.databases = Databases(self)
         self.tables = Tables(self)
         self.webhooks = Webhooks(self)
+        self.data_acceleration_report = DataAccelerationReport(self)
         self._namespace = Namespace()
 
         if use_server_version: